Polishing scripts

+++ install.sh
@@ -8,16 +8,12 @@ SYMBOLIC_ICON_FOLDER=$HOME/.local/share/icons/hicolor/symbolic/apps
 RAMDISK_FOLDER=$HOME/RAMDisk
 RAMDISK_PERSISTENT_FOLDER=$HOME/.RAMDisk
 
-# Create folder to mount ramfs
-# mkdir -p $RAMDISK_FOLDER
-
 # Create entry in /etc/fstab
-# RAMDisk is not mounted automatically, user service will mount at login
-# https://forums.linuxmint.com/viewtopic.php?t=123354
+echo "INFO: Creating ramfs entry in /etc/fstab."
 sudo sed -i "/#ramdisk-$USER/d" /etc/fstab
 echo "none $RAMDISK_FOLDER ramfs rw,exec,noauto,user,mode=1777 0 0 #ramdisk-$USER" | sudo tee -a /etc/fstab
 
-# Update fstab
+# Make `mount` recognize changes in `fstab`: https://unix.stackexchange.com/q/477794/340672
 sudo systemctl daemon-reload
 
 # Copy startup and shutdown scripts
@@ -25,18 +21,21 @@ sudo systemctl daemon-reload
 # https://osric.com/chris/accidental-developer/2019/01/cp-mv-ownership-attributes/
 mkdir -p $BIN_FOLDER
 
-# Need exec permissions?
+# Check which RAMDisk scripts to install - with or without executable permissions on the drive?
 while true; do
-read -p "Do you need executable permissions on the RAMDisk? (yes/no) " yn
+read -p "QUESTION: Do you need executable permissions on the RAMDisk? (yes/no) " yn
 case $yn in 
-    yes ) echo ok;
+    yes )
+        echo -e "INFO: Ok. Copying startup and shutdown scripts."
         cp ./src/tiny-ramdisk-ramfs-startup-exec.sh $BIN_FOLDER/tiny-ramdisk-ramfs-startup.sh
         cp ./src/tiny-ramdisk-ramfs-shutdown-exec.sh $BIN_FOLDER/tiny-ramdisk-ramfs-shutdown.sh
+        echo -e "INFO: Installing the policy file."
         sudo cp ./src/com.tiny-ramdisk.policy /usr/share/polkit-1/actions
         sudo mv /usr/share/polkit-1/actions/com.tiny-ramdisk.policy /usr/share/polkit-1/actions/com.$USER.tiny-ramdisk.policy
         sudo sed -i "s;%USER%;$USER;g" /usr/share/polkit-1/actions/com.$USER.tiny-ramdisk.policy
         break;;
-    no ) echo ok;
+    no )
+        echo -e "INFO: Ok. Copying startup and shutdown scripts."
         cp ./src/tiny-ramdisk-ramfs-startup-noexec.sh $BIN_FOLDER/tiny-ramdisk-ramfs-startup.sh
         cp ./src/tiny-ramdisk-ramfs-shutdown-noexec.sh $BIN_FOLDER/tiny-ramdisk-ramfs-shutdown.sh
         break;;
@@ -44,9 +43,11 @@ case $yn in
 esac
 done
 
+# Add executable permissions to RAMDisk scripts
 chmod +x $BIN_FOLDER/tiny-ramdisk-ramfs*
 
 # Install icons for notifications and pkexec dialog (if needed)
+echo "INFO: Installing application icons."
 mkdir -p $ICON_FOLDER
 mkdir -p $SYMBOLIC_ICON_FOLDER
 cp ./resources/com.petrstepanov.TinyRAMDisk.svg $ICON_FOLDER
@@ -54,6 +55,7 @@ cp ./resources/com.petrstepanov.TinyRAMDisk-symbolic.svg $SYMBOLIC_ICON_FOLDER
 xdg-desktop-menu forceupdate
 
 # Create folder for user systemd scripts
+echo "INFO: Installing systemd service."
 mkdir -p $SYSTEMD_FOLDER
 
 # Copy systemd services
@@ -66,9 +68,9 @@ systemctl --user disable tiny-ramdisk-ramfs
 systemctl --user daemon-reload
 
 systemctl --user enable tiny-ramdisk-ramfs
-echo Waiting for the service to start...
+echo "INFO: Starting the RAMDisk service..."
 systemctl --user start tiny-ramdisk-ramfs
 
-# Open ramdisk folder in Files
-echo "Path to your RAMDisk is $RAMDISK_FOLDER"
+# Show RAMDisk folder in Files application
+echo "INFO: Done! Path to your RAMDisk is ${RAMDISK_FOLDER}"
 xdg-open $RAMDISK_FOLDER
